+++ src/Core/Core/Execution/Utilities/VariableValueBuilder.cs
@@ -186,7 +186,7 @@ namespace HotChocolate.Execution
             return value;
         }
 
-        private void CheckForInvalidValueType(
+        private static void CheckForInvalidValueType(
             VariableDefinitionNode variableDefinition,
             Variable variable)
         {
@@ -203,7 +203,7 @@ namespace HotChocolate.Execution
             }
         }
 
-        private void CheckForInvalidValueType(
+        private static void CheckForInvalidValueType(
             VariableDefinitionNode variableDefinition,
             Variable variable,
             IValueNode value)