+++ src/Core/Language/Visitors/SchemaSyntaxSerializer.cs
@@ -373,7 +373,7 @@ namespace HotChocolate.Language
             }
         }
 
-        private void WriteDirectives(
+        private static void WriteDirectives(
             IReadOnlyList<DirectiveNode> directives,
             DocumentWriter writer)
         {