+++ src/Core/Subscriptions.InMemory/InMemoryEventStream.cs
@@ -89,6 +89,7 @@ namespace HotChocolate.Subscriptions
         public void Dispose()
         {
             Dispose(true);
+            GC.SuppressFinalize(this);
         }
 
         #endregion