+++ CHANGELOG.md
@@ -36,7 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Includes directive definitions in serialized schema [#717](https://github.com/ChilliCream/hotchocolate/issues/717)
 - Field types are now validated. [#713](https://github.com/ChilliCream/hotchocolate/issues/713)
-
+- Variables in object values and lists are now correctly recognised [#215](https://github.com/ChilliCream/hotchocolate/issues/215) and [#745](https://github.com/ChilliCream/hotchocolate/issues/745).
 
 ## [0.8.2] - 2019-04-10